# table: make better use of already allocated memory for snap compression