@syndtr 2021-08-19
My understanding is that `make([]byte, n)` would always returns slice with `len(n) ==  cap(n)` and since `w.compressionScratch` itself never resized in the first place, aren't checking only the length should be enough?